backend/requirements.txt
chromadb==1.5.6
fastapi==0.135.3
langchain_community==0.4.1
langchain_core==1.2.26
langchain_groq==1.1.2
langchain_text_splitters==1.1.1
numpy==2.4.4
pydantic==2.12.5
python-dotenv==1.2.2
sentence_transformers==5.3.0
